requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Psycopg2==2.6.1
sqlalchemy==1.0.12
wtforms==2.0.2
flask-wtf==0.12
requests==2.9.1
wtforms-alchemy==0.16.0
passlib==1.6.5
flask-pagedown==0.2.1
flask-markdown==0.3
flask-bootstrap==3.3.5.7
markupsafe==0.23
flask-nav==0.5
flask-security==1.7.5
bcrypt
requests-cache
pillow
